implement gemma3 fp8 encoder with layerwise casting and pinned memory, add model manager caching for transformer state_dict and vae persistence, and update generation pipeline schemas for stg and lora scaling

+++ src/llm_dit/config.py
@@ -425,7 +425,7 @@ class LTX2Config:
     # Quantization
     quantize: str = "fp8"  # Transformer quantization (none, fp8)
     skip_cleanup: bool = False  # Skip memory cleanup between stages
-    gemma_variant: str = "bf16"  # Gemma3 backbone: bf16, 8bit, q4-qat
+    gemma_variant: str = "fp8"  # Gemma3 backbone: bf16, fp8, 8bit, q4-qat
 
     # Preset configuration
     default_preset: str = ""  # Default preset to load (e.g., "cinematic")
